requirements.txt
firecrawl-py
groq
python-dotenv
typer
rich
tqdm
langgraph
pydantic